pyproject.toml
[tool.poetry]
name = "repl_python3_beta_waifu-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
Flask = "^2.0.1"
"dislash.py" = "^1.4.9"
airi = "^0.1.1"
gTTS = "^2.2.3"
orjson = "^3.6.6"
typing-extensions = "^4.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
